==> CODE/16_weather_download.ipynb <==
import pandas as pd
import numpy as np
import os.path
import requests
import json

root_path = os.path.dirname(os.getcwd())

# Import food inspection data
inspections = pd.read_csv(os.path.join(root_path, "DATA/food_inspections.csv"))

# Create weather dataframe with entry for each inspection
weather = pd.DataFrame({
    "inspection_id": inspections.inspection_id,
    "date": inspections.inspection_date,
})

# Create list of weather data columns
features = [
    "precip_intensity",
    "temperature_max",
    "wind_speed",
    "humidity"
]

# Add empty columns
weather = pd.concat([weather, pd.DataFrame(columns=features)])

# Check if weather data exists, import or start empty dataframe
if os.path.isfile(os.path.join(root_path, "DATA/weather.csv")):
    print("records found")
    old_weather = pd.read_csv(os.path.join(root_path, "DATA/weather.csv"))
else:
    old_weather = pd.DataFrame(columns=weather.columns)

# Register for a free key at https://darksky.net/dev/register
KEY = "c30a95e4e9f7f2e90966361288918b9e"

# Check if weather record can be found in old data
# If not,request record from darksky.net
# All records pertain to a single location in central Chicago

limit_reached = False

def get_weather(group):
    global limit_reached
    
    date = group.iloc[0].date
    record = old_weather[old_weather.date == date]
    if (not record.empty) & (not record.isnull().values.any()):
        print(date, "record found")
        group.loc[:, features] = record.iloc[0][features].values
        
    elif not limit_reached:
        url = "https://api.darksky.net/forecast/%s/41.836944,-87.684722,%s" % (KEY, date)
        response = requests.get(url)
        
        if response.status_code == 200:
            print(date, "request successful")
            data = response.json()["daily"]["data"][0] 
            group[features] = [
                data["precipIntensity"],
                data["temperatureMax"],
                data["windSpeed"],
                data["humidity"]
            ]
        elif response.status_code == 403:
            print("request limit reached")
            limit_reached = True
        else:
            print(response)
    return group

# Time consuming
weather = weather.groupby(by="date").apply(get_weather)

# Check if process is complete
if weather.isnull().values.any():
    print("Process incomplete, records missing")
else:
    print("Process complete, all records obtained")

Darksky.net limits free accounts to 1000 calls per day. If records are missing, save the result and run the notebook from the top after waiting 24 hours, upgrading your account or obtaining a new API key.

# Save result
weather.to_csv(os.path.join(root_path, "DATA/weather.csv"), index=False)
